function_app/requirements.txt
azure-functions
azure-identity
azure-mgmt-containerinstance
requests
azure-core
azure-common
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
requests>=2.31.0